+++ Jint/Runtime/Interpreter/Statements/JintVariableDeclaration.cs
@@ -14,8 +14,7 @@ namespace Jint.Runtime.Interpreter.Statements
         private sealed class ResolvedDeclaration
         {
             internal JintExpression Left;
-            internal ArrayPattern LeftArrayPattern;
-            internal ObjectPattern LeftObjectPattern;
+            internal BindingPattern LeftPattern;
             internal JintExpression Init;
             internal JintIdentifierExpression LeftIdentifier;
             internal bool EvalOrArguments;
@@ -35,34 +34,26 @@ namespace Jint.Runtime.Interpreter.Statements
 
                 JintExpression left = null;
                 JintExpression init = null;
-                ArrayPattern leftArrayPattern = null;
-                ObjectPattern leftObjectPattern = null;
+                BindingPattern bindingPattern = null;
                 if (declaration.Init != null)
                 {
                     if (declaration.Id is Expression expression)
                     {
                         left = JintExpression.Build(_engine, expression);
                     }
-                    else if (declaration.Id is ArrayPattern arrayPattern)
+                    else if (declaration.Id is BindingPattern bp)
                     {
-                        leftArrayPattern = arrayPattern;
-                    }
-                    else if (declaration.Id is ObjectPattern objectPattern)
-                    {
-                        leftObjectPattern = objectPattern;
+                        bindingPattern = bp;
                     }
                         
-                    init = declaration.Init != null
-                        ? JintExpression.Build(_engine, declaration.Init)
-                        : null;
+                    init = JintExpression.Build(_engine, declaration.Init);
                 }
 
                 var leftIdentifier = left as JintIdentifierExpression;
                 _declarations[i] = new ResolvedDeclaration
                 {
                     Left = left,
-                    LeftArrayPattern = leftArrayPattern,
-                    LeftObjectPattern = leftObjectPattern,
+                    LeftPattern = bindingPattern,
                     LeftIdentifier = leftIdentifier,
                     EvalOrArguments = leftIdentifier?._expressionName == "eval" || leftIdentifier?._expressionName == "arguments",
                     Init = init
@@ -78,18 +69,11 @@ namespace Jint.Runtime.Interpreter.Statements
                 var declaration = declarations[i];
                 if (declaration.Init != null)
                 {
-                    if (declaration.LeftArrayPattern != null)
+                    if (declaration.LeftPattern != null)
                     {
-                        JintAssignmentExpression.ArrayPatternAssignmentExpression.AssignToPattern(
+                        BindingPatternAssignmentExpression.ProcessPatterns(
                             _engine,
-                            declaration.LeftArrayPattern,
-                            declaration.Init.GetValue());
-                    }
-                    else if (declaration.LeftObjectPattern != null)
-                    {
-                        JintAssignmentExpression.ObjectPatternAssignmentExpression.AssignToPattern(
-                            _engine,
-                            declaration.LeftObjectPattern,
+                            declaration.LeftPattern,
                             declaration.Init.GetValue());
                     }
                     else if (declaration.LeftIdentifier == null
